`std::execution` is hidden to `.cpp`
TBB 2021 (included with `std::execution` with Qt leads to macro
collision.
Bot TBB and Qt have `emit` macro.
On Linux this lead to error.
See:
https://forum.qt.io/topic/122225/trying-to-use-c-17-parallel-algorithms-
with-qt/9

+++ src/h5geo/h5core_segy.cpp
@@ -392,7 +392,7 @@ bool readSEGYTraces(
       } else if (format == h5geo::SegyFormat::FourByte_integer) {
         for (size_t i = 0; i < nSamp; i++) {
           TRACE(i, j) =
-              to_native_endian(m_int[j * bytesPerTrc / 4 + i + 60], endian);
+              (float)to_native_endian(m_int[j * bytesPerTrc / 4 + i + 60], endian);
         }
       } else if (format == h5geo::SegyFormat::FourByte_IEEE) {
         for (size_t i = 0; i < nSamp; i++) {
